+++ src/services/handlers/PublishDraftHandler.php
@@ -13,9 +13,7 @@ use Craft;
 use craft\base\ElementInterface;
 use craft\errors\InvalidElementException;
 use craft\services\Drafts as DraftRepository;
-use lilthq\craftliltplugin\parameters\CraftliltpluginParameters;
 use lilthq\craftliltplugin\records\SettingRecord;
-use lilthq\craftliltplugin\records\TranslationRecord;
 use Throwable;
 use yii\base\Exception;
 
@@ -39,90 +37,6 @@ class PublishDraftHandler
 
         if (!$draftElement) {
             return;
-        }
-
-        // Merge canonical changes for supertable fields in current draft before publishing
-        if (
-            class_exists('verbb\supertable\SuperTable')
-            || class_exists('benf\neo\Plugin')
-        ) {
-            $translation = TranslationRecord::findOne(['translatedDraftId' => $draftId]);
-            $translations = TranslationRecord::findAll(
-                [
-                    'jobId' => $translation->jobId,
-                    'status' => TranslationRecord::STATUS_PUBLISHED
-                ]
-            );
-
-            foreach ($translations as $translation) {
-                $draftElementLanguageToUpdate = Craft::$app->elements->getElementById(
-                    $draftId,
-                    null,
-                    $translation->targetSiteId
-                );
-                $draftElementLanguageToUpdate->mergingCanonicalChanges = true;
-
-                $fieldLayout = $draftElementLanguageToUpdate->getFieldLayout();
-                $fields = $fieldLayout ? $fieldLayout->getCustomFields() : [];
-                foreach ($fields as $field) {
-                    // Check if the field is of Super Table type and the required classes and methods are available
-                    if (
-                        get_class($field) === CraftliltpluginParameters::CRAFT_FIELDS_SUPER_TABLE
-                        && method_exists('verbb\supertable\SuperTable', 'getService')
-                        && method_exists('verbb\supertable\SuperTable', 'getInstance')
-                    ) {
-                        // Get the Super Table plugin instance
-                        $superTablePluginInstance = call_user_func(['verbb\supertable\SuperTable', 'getInstance']);
-
-                        // Get the Super Table plugin service
-                        /** @var \verbb\supertable\services\Service $superTablePluginService */
-                        $superTablePluginService = $superTablePluginInstance->getService();
-
-                        // Duplicate the blocks for the field
-                        $superTablePluginService->duplicateBlocks(
-                            $field,
-                            $draftElementLanguageToUpdate->getCanonical(),
-                            $draftElementLanguageToUpdate
-                        );
-
-                        continue;
-                    }
-
-                    if (
-                        get_class($field) === CraftliltpluginParameters::BENF_NEO_FIELD
-                        && class_exists('benf\neo\Plugin')
-                        && method_exists('benf\neo\Plugin', 'getInstance')
-                    ) {
-                        // Get the Neo plugin instance
-                        /** @var \benf\neo\Plugin $neoPluginInstance */
-                        $neoPluginInstance = call_user_func(['benf\neo\Plugin', 'getInstance']);
-
-                        // Get the Neo plugin Fields service
-                        /** @var \benf\neo\services\Fields $neoPluginFieldsService  */
-                        $neoPluginFieldsService = $neoPluginInstance->get('fields');
-
-                        // Clear current neo field value
-                        $neoField = $draftElementLanguageToUpdate->getFieldValue($field->handle);
-                        foreach ($neoField as $block) {
-                            if (!$block instanceof ElementInterface) {
-                                continue;
-                            }
-
-                            Craft::$app->getElements()->deleteElement($block);
-                        }
-                        Craft::$app->getElements()->saveElement($draftElementLanguageToUpdate);
-
-                        // Duplicate the blocks for the field
-                        $neoPluginFieldsService->duplicateBlocks(
-                            $field,
-                            $draftElementLanguageToUpdate->getCanonical(),
-                            $draftElementLanguageToUpdate
-                        );
-
-                        continue;
-                    }
-                }
-            }
         }
 
         $enableEntriesForTargetSitesRecord = SettingRecord::findOne(['name' => 'enable_entries_for_target_sites']);