+++ tests/integration/modules/FetchTranslationFromConnectorCest.php
@@ -43,10 +43,8 @@ class FetchTranslationFromConnectorCest extends AbstractIntegrationCest
      * @return void
      * @throws InvalidFieldException
      */
-    public function testExecuteInstantSuccess(IntegrationTester $I, $scenario): void
+    public function testExecuteInstantSuccess(IntegrationTester $I): void
     {
-        $scenario->skip('Content is not getting updated and missing in source content');
-
         Db::truncateTable(Craft::$app->queue->tableName);
 
         $user = Craft::$app->getUsers()->getUserById(1);
@@ -187,10 +185,8 @@ class FetchTranslationFromConnectorCest extends AbstractIntegrationCest
      * @return void
      * @throws InvalidFieldException
      */
-    public function testExecuteVerifiedSuccess(IntegrationTester $I, $scenario): void
+    public function testExecuteVerifiedSuccess(IntegrationTester $I): void
     {
-        $scenario->skip('Content is not getting updated and missing in source content');
-
         Db::truncateTable(Craft::$app->queue->tableName);
 
         $user = Craft::$app->getUsers()->getUserById(1);