+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 4.2.2 - 2023-09-22
+### Fixed
+- Empty user id on draft creation
+- Nested blocks getting removed on current entry version when translations downloaded
+
 ## 4.2.1 - 2023-09-05
 ### Fixed
 - Download diagnostic data functionality